Revert "API: Return URLs in escaped (percent-encoded) form"
This reverts commit b79b4677f910d101f8f392fb1ec2618ddb4c38da.

This breaks the API in a big way, as it mangles RTMP URLs (spaces are now
%20), and might even wrongly escape some characters that didn't need to
be. This is a major API break as URLs returned are now incompatible.

Applications should instead run:
setlocale (LC_ALL, "");
and expect the environment to be in UTF-8.

Applications should also expect all the URLs and other strings from libquvi to
be in UTF-8 encoding.